requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.5.1
google==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==4.7.6
praw==7.1.0
prawcore==1.5.0
python-dotenv==0.15.0
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.5.1
